# src/main/resources/api/openapi.yml
openapi: 3.0.2
info:
  title: Cena - Menu generation
  version: '1'
  description: Cena provides a REST API for menu generation
  contact:
    name: Alexandre Carbenay
  license:
    name: GNU GENERAL PUBLIC LICENSE
    url: 'http://www.gnu.org/licenses/'
servers:
  - description: Production server
    url: 'https://cena.adhuc.org/api'
paths:
...
  /ingredients:
    get:
      summary: Get the ingredients collection. Ingredients collection is accessible for every user, either authenticated or not
      operationId: listIngredients
      tags:
        - Ingredients
      responses:
        200:
          $ref: '#/components/responses/Ingredients'
    post:
      summary: Create an ingredient in ingredients collection. The specified ingredient name must not already be used by an existing ingredient. Only an ingredient manager can create ingredients
      operationId: createIngredient
      tags:
        - Ingredients
      security:
        - basicAuth:[]
      requestBody:
        $ref: '#/components/requestBodies/CreateIngredient'
      responses:
        201:
          $ref: '#/components/responses/IngredientCreated'
        400:
          $ref: '#/components/responses/400-BadRequest'
        401:
          $ref: '#/components/responses/401-Unauthorized'
        403:
          $ref: '#/components/responses/403-Forbidden'
    delete:
      summary: Delete all ingredients from the system. Only a super administrator can delete all ingredients directly. An ingredient manager can delete specific ingredients through the corresponding ingredient resource.
      operationId: deleteIngredients
      tags:
        - Ingredients
      security:
        - basicAuth:[]
      responses:
        204:
          description: All ingredients have been deleted from the system
        401:
          $ref: '#/components/responses/401-Unauthorized'
        403:
          $ref: '#/components/responses/403-Forbidden'
...
  /menus:
    post:
      summary: Create a menu in menus collection for the currently authenticated user
      operationId: createMenu
      tags:
        - Menus
      security:
        - basicAuth:[]
      requestBody:
        $ref: '#/components/requestBodies/CreateMenu'
      responses:
        201:
          $ref: '#/components/responses/MenuCreated'
        400:
          $ref: '#/components/responses/400-BadRequest'
        401:
          $ref: '#/components/responses/401-Unauthorized'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
...
  requestBodies:
    CreateIngredient:
      required: true
      description: Ingredient creation request
      content:
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Ingredient'
        application/json:
          schema:
            $ref: '#/components/schemas/Ingredient'
...
    CreateMenu:
      required: true
      description: Menu creation request
      content:
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Menu'
        application/json:
          schema:
            $ref: '#/components/schemas/Menu'
  responses:
...
    Ingredients:
      description: Collection containing all the ingredients found in the system. For now, this collection is not sortable nor pageable, but those behaviors may be implemented in the future.
      content:
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Ingredients'
        application/json:
          schema:
            $ref: '#/components/schemas/Ingredients'
    IngredientCreated:
      description: Successful ingredient creation response
      headers:
        Location:
          description: Link to the created ingredient. Can be used to retrieve, update or delete the created ingredient
          schema:
            type: string
            example: 'https://cena.adhuc.org/api/ingredients/3fa85f64-5717-4562-b3fc-2c963f66afa6'
...
    MenuCreated:
      description: Successful menu creation response
      headers:
        Location:
          description: Link to the created menu. Can be used to retrieve, update or delete the created menu
          schema:
            type: string
            example: 'https://cena.adhuc.org/api/menus/07549d0c-b8ef-40a0-af3b-5b6107706175'
    400-BadRequest:
      description: Bad syntax or incorrect request
      content:
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Error'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    401-Unauthorized:
      description: An authentication is required
      content:
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Error'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    403-Forbidden:
      description: Authenticated user does not have access to resource
      content:
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Error'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Ingredients:
      description: A collection of ingredients, containing ingredients detail for each ingredient
      type: object
      required:
        - '_links'
      properties:
        '_embedded':
          type: object
          readOnly: true
          required:
            - data
          properties:
            data:
              description: The ingredients collection data, containing ingredients detail
              type: array
              items:
                $ref: '#/components/schemas/Ingredient'
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this ingredients collection. Can be used to create new ingredients in the collection
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://cena.adhuc.org/api/ingredients'
    Ingredient:
      description: An ingredient detail, containing its identity, name and measurement types
      type: object
      required:
        - id
        - name
        - '_links'
      properties:
        id:
          description: Ingredient identity
          type: string
          format: uuid
          readOnly: true
          example: '3fa85f64-5717-4562-b3fc-2c963f66afa6'
        name:
          description: Ingredient name, unique in the system
          type: string
          minLength: 1
          example: 'Tomato'
        measurementTypes:
          description: Ingredient measurement types, used to determine which units of measurements ingredients accept when associated to recipes
          type: array
          items:
            type: string
            enum:
              - WEIGHT
              - VOLUME
              - COUNT
              - AT_CONVENIENCE
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this ingredient detail. Can be used to update or delete the ingredient
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://cena.adhuc.org/api/ingredients/3fa85f64-5717-4562-b3fc-2c963f66afa6'
      additionalProperties: true
...
    RecipeId:
      description: Recipe identity
      type: string
      format: uuid
      readOnly: true
      example: 'd71e2fc7-09e3-4241-97a5-dc3383d35e98'
...
    Menu:
      description: A menu detail
      type: object
      required:
        - date
        - mealType
        - covers
        - mainCourseRecipes
        - '_links'
      properties:
        date:
          description: Menu date
          type: string
          format: date
          example: '2020-01-01'
        mealType:
          $ref: '#/components/schemas/MealType'
        covers:
          description: Number of covers of a menu
          type: integer
          example: 2
          minimum: 1
        mainCourseRecipes:
          description: List of recipe identities related to menu's main course. Main course must consist of at least one recipe
          type: array
          items:
            $ref: '#/components/schemas/RecipeId'
          minItems: 1
        '_links':
          type: object
          readOnly: true
          required:
            - self
            - mainCourseRecipes
          properties:
            self:
              description: Link to this menu detail. Can be used to update or delete the menu
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://cena.adhuc.org/api/menus/2020-01-01_LUNCH'
            mainCourseRecipes:
              description: Links to this menu's main course recipes
              type: array
              items:
                type: object
                required:
                  - href
                properties:
                  href:
                    type: string
                    example: 'https://cena.adhuc.org/api/recipes/d71e2fc7-09e3-4241-97a5-dc3383d35e98'
      additionalProperties: true
    MealType:
      description: Meal type for a menu
      type: string
      enum: ['LUNCH', 'DINER']
    Error:
      description: General error structure
      type: object
      properties:
        timestamp:
          description: The date this error occurred on the server
          type: string
          format: date-time
        status:
          description: The HTTP status code applicable to this error
          type: number
          example: 400
        error:
          description: A human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the error
          type: string
          example: Bad request
        message:
          description: A human-readable explanation specific to this occurrence of the error
          type: string
          example: Ingredient name 'Tomato' is already used
        path:
          description: The path used to access to the resource returning the error
          type: string
          example: '/api/ingredients'
        code:
          description: An application-specific error code
          type: number
          example: 900100